+++ src/views/index/user/index.vue
@@ -45,7 +45,7 @@
         </div>
         <footer>
           <span class="clickable" @click="$router.push({path: `/index/`})">全部星球</span>
-          <span class="clickable" @click="$router.push({path: `/login/`})">创建星球</span>
+          <span class="clickable" @click="$router.push({path: `/group/create`})">创建星球</span>
         </footer>
       </aside>
     </div>
